requirements.txt
Flask==1.1.2
MySQL-python==8.0.24